Add files via upload
The user needs to wear a sensor embedded glove to control a robotic arm. Contains the arduino code for the glove designed with IMU sensors placed in two locations which provide signal captured from the 3 dimensional motion of the user's arm. The data is translated to actuate the robotic arm's servo motors. This acts are a user interface and this  is an initial step towards implementation of AI (imitative learning) in the 3DOF robotic arm.